VST2Plugin: buffer incoming MIDI messages
some MIDI plugins want the 'VstEvent' pointers to be valid for the whole duration of the process funtion.
we store incoming MIDI and SysEx messages in a buffer, send them over in one go before the process routine and clear the buffer afterwards.
'VstEvents' - the array holding the VstEvent pointers - is a member of VST2Plugin and grows as needed.