VstPluginUGen: fix possible crasher bug.
because of a bug(?) in SuperCollider, UGen commands might be sent before the constructor is called! This can lead to various crashes when member functions are called.
my workaround is to compare an instance member against a magic number. if the two match, it means the constructor has been called.