make sure to load/save presets and reset the plugin on the UI thread
the reason is to avoid race conditions with concurrent UI updates.
of course, this only applies if we have a plugin UI.
In Supercollider, we always force "async=true" for plugin UIs, otherwise we can't defer to the UI thread in a real-time safe way.
In Pd, we just just defer, because it's more important to keep the synchronous behavior.
Finally, we simplify the plugin loading code by making sure that we only try to open the plugin on the UI thread if a) the plugin has an editor and b) the host has an event loop