vst: don't use multi-point parameter automation
it takes a lot of memory and few VST3 plugins actually support it and Pd resp. SuperCollider generally operate on a small blocksize.
This means, there can only be a single (sample-accurate) parameter point per processing block.
The old, mulit-point parameter automation is still available as a compile time option.
LATER make it a runtime option?