sc/VSTPlugin: improve RT thread checks
Instead of saving the thread ID of the current RT thread, we simply set a thread local flag in the constructor.
For supernova we also need to do this in the next routine and before opening a plugin (= the first unit command)
In the IPluginListener callbacks we can then simply check the thread local flag to see if we're being called from a RT thread.

This simpler and faster than calling std::this_thread::get_id() before every function that might cause a plugin callback.