X11: fix/improve event loop
Windows, timers and event handlers can only be added/removed on the UI thread;
They are pushed to a dedicated queue, so that we can't accidentally modify the
timer list while we are iterating over it.
Poll functions are added/removed via commands, so we don't need any synchronization.
We only need a mutex for the command queue itself.