PluginServer: when reading a preset, don't send parameter automation events to UI threads
The UI receive queue is rather small and meant for automating parameters manually in the UI,
it can overflow easily if the plugin has many parameters.
(The client is already blocking in the UI thread and thus cannot drain the queue.)
When the queue overflows, not only do we lose parameter data,
we may also stall for a long time because of our overflow protection.

As a workaround, we use a dedicated queue for parameter changes during preset loading;
after the preset has finished loading, we drain the queue and send the data through the NRT channel.