sc: correctly handle /vst_param replies with deferred processing
if plugin processing is deferred (threaded or sandboxed/bridged), the true parameter value or display is only available after the process() function.
In this case, when the user sets a parameter with the /set command, we set the corresponding bit in the parameter bitset.
After we process the plugin, we check the bitset and retrieve the corresponding parameters, so we can send /vst_param messages to the client.
For multithreaded plugins we use double buffering to accomodate the additional one block delay.

Also, we only unset 'isSettingParam_' at the end of the next function, so we can handle outgoing parameter changes for VST3 plugins or threaded/sandboxed/bridged plugins.