x11: use proper priority queue for timers
... so we can easily add/remove timers recursively, i.e. no need for 'newTimers_' or 'invalidate'.
We just need to make sure to pop the timer before rescheduling and calling it!
NB: we must reschedule the timer *before* calling it, so it can be found and removed from within the timer function!

Also, for the sake of both simplicity and consistency, let's use a single timer for all poll functions, just like in the Windows and macOS implementation.